requirements.txt
google-generativeai>=0.5.0
gradio>=4.0.0
Pillow